Fix output bound calculations (#168)
This PR is to ensure that the bounds on unscaled output vars are
calculated from unscaled input bounds. Likewise, it ensures that bounds
on scaled output vars are calculated from scaled input bounds

**Legal Acknowledgement**\
By contributing to this software project, I agree my contributions are
submitted under the BSD license.
I represent I am authorized to make the contributions and grant the
license.
If my employer has rights to intellectual property that includes these
contributions,
I represent that I have received permission to make contributions and
grant the required license on behalf of that employer.