Patch issue 173 (#175)
This path is to fix the bug found in issue 173. If a user passes a
scaler and scaled input bounds when creating a LinearTreeDefinition,
unscaled input bounds will be calculated.

If only scaled input bounds are provided, unscaled input bounds will
remain None.

No output bounds will be assigned to unscaled output vars if unscaled
input bounds is None.

**Legal Acknowledgement**\
By contributing to this software project, I agree my contributions are
submitted under the BSD license.
I represent I am authorized to make the contributions and grant the
license.
If my employer has rights to intellectual property that includes these
contributions,
I represent that I have received permission to make contributions and
grant the required license on behalf of that employer.